finsihed 2 base cruds

+++ react-vite/src/components/GridTool/GridTool.jsx
@@ -14,12 +14,78 @@ const GridTool = ({ symbols, onEditSymbol, onDeleteSymbol }) => {
   const [offsetY, setOffsetY] = useState(0);
   const [selectedSymbol, setSelectedSymbol] = useState(null);
   const symbolsArr = Object.values(symbols);
-//   const [fSymbols, setFSymbols] = useState(symbols);
+
+  const symbolBoundsCheck = (symbol, mouseX, mouseY) => {
+    switch (symbol.type) {
+        case 'Terminal':
+          return (
+            mouseX >= symbol.x_position - SYMBOL_SIZE / 2 &&
+            mouseX <= symbol.x_position + SYMBOL_SIZE / 2 &&
+            mouseY >= symbol.y_position - SYMBOL_SIZE / 4 &&
+            mouseY <= symbol.y_position + SYMBOL_SIZE / 4
+          );
+        case 'Decision':
+          return (
+            mouseX >= symbol.x_position - SYMBOL_SIZE / 1.75 &&
+            mouseX <= symbol.x_position + SYMBOL_SIZE / 1.75 &&
+            mouseY >= symbol.y_position - SYMBOL_SIZE / 3 &&
+            mouseY <= symbol.y_position + SYMBOL_SIZE / 3
+          );
+        case 'Process':
+          return (
+            mouseX >= symbol.x_position - SYMBOL_SIZE / 2 &&
+            mouseX <= symbol.x_position + SYMBOL_SIZE / 2 &&
+            mouseY >= symbol.y_position - SYMBOL_SIZE / 4 &&
+            mouseY <= symbol.y_position + SYMBOL_SIZE / 4
+          );
+        case 'Data':
+          return (
+            mouseX >= symbol.x_position - SYMBOL_SIZE / 2 &&
+            mouseX <= symbol.x_position + SYMBOL_SIZE / 1.33 &&
+            mouseY >= symbol.y_position - SYMBOL_SIZE / 4 &&
+            mouseY <= symbol.y_position + SYMBOL_SIZE / 4
+          );
+        default:
+          return false;
+      }
+}
 
   // eslint-disable-next-line react-hooks/exhaustive-deps
   useEffect(() => {
         const canvas = canvasRef.current;
         const ctx = canvas.getContext('2d');
+
+    const drawSelectedSymbol = (ctx, type, x, y) => {
+        switch (type) {
+            case 'Terminal':
+                ctx.ellipse(x, y, SYMBOL_SIZE / 2, SYMBOL_SIZE / 4, 0, 0, Math.PI * 2);
+                ctx.stroke();
+                break;
+            case 'Decision':
+                ctx.beginPath();
+                ctx.moveTo(x, y - SYMBOL_SIZE / 3);
+                ctx.lineTo(x + SYMBOL_SIZE / 1.75, y);
+                ctx.lineTo(x, y + SYMBOL_SIZE / 3);
+                ctx.lineTo(x - SYMBOL_SIZE / 1.75, y);
+                ctx.closePath();
+                ctx.stroke();
+                break;
+            case 'Process':
+                ctx.strokeRect(x - SYMBOL_SIZE / 2, y - SYMBOL_SIZE / 4, SYMBOL_SIZE, SYMBOL_SIZE / 2);
+                break;
+            case 'Data':
+                ctx.beginPath();
+                ctx.moveTo(x - SYMBOL_SIZE / 2, y - SYMBOL_SIZE / 4);
+                ctx.lineTo(x + SYMBOL_SIZE / 2, y - SYMBOL_SIZE / 4);
+                ctx.lineTo(x + SYMBOL_SIZE / 1.33, y + SYMBOL_SIZE / 4);
+                ctx.lineTo(x - SYMBOL_SIZE / 4, y + SYMBOL_SIZE / 4);
+                ctx.closePath();
+                ctx.stroke();
+                break;
+            default:
+                break;
+        }
+    }
 
     const resizeCanvas = () => {
         const clientWidth = canvas.clientWidth;
@@ -65,35 +131,7 @@ const GridTool = ({ symbols, onEditSymbol, onDeleteSymbol }) => {
                     ctx.strokeStyle = 'yellow';
                     ctx.lineWidth = 4;
 
-                    switch (type) {
-                        case 'Terminal':
-                            ctx.ellipse(x_position, y_position, SYMBOL_SIZE / 2, SYMBOL_SIZE / 4, 0, 0, Math.PI * 2);
-                            ctx.stroke();
-                            break;
-                        case 'Decision':
-                            ctx.beginPath();
-                            ctx.moveTo(x_position, y_position - SYMBOL_SIZE / 3);
-                            ctx.lineTo(x_position + SYMBOL_SIZE / 1.75, y_position);
-                            ctx.lineTo(x_position, y_position + SYMBOL_SIZE / 3);
-                            ctx.lineTo(x_position - SYMBOL_SIZE / 1.75, y_position);
-                            ctx.closePath();
-                            ctx.stroke();
-                            break;
-                        case 'Process':
-                            ctx.strokeRect(x_position - SYMBOL_SIZE / 2, y_position - SYMBOL_SIZE / 4, SYMBOL_SIZE, SYMBOL_SIZE / 2);
-                            break;
-                        case 'Data':
-                            ctx.beginPath();
-                            ctx.moveTo(x_position - SYMBOL_SIZE / 2, y_position - SYMBOL_SIZE / 4);
-                            ctx.lineTo(x_position + SYMBOL_SIZE / 2, y_position - SYMBOL_SIZE / 4);
-                            ctx.lineTo(x_position + SYMBOL_SIZE / 1.33, y_position + SYMBOL_SIZE / 4);
-                            ctx.lineTo(x_position - SYMBOL_SIZE / 4, y_position + SYMBOL_SIZE / 4);
-                            ctx.closePath();
-                            ctx.stroke();
-                            break;
-                        default:
-                            break;
-                    }
+                    drawSelectedSymbol(ctx, type, x_position, y_position);
                 }
 
             });
@@ -106,49 +144,16 @@ const GridTool = ({ symbols, onEditSymbol, onDeleteSymbol }) => {
         const mouseX = e.clientX - rect.left;
         const mouseY = e.clientY - rect.top;
 
-        const clickedSymbol = symbolsArr.find(symbol => {
-            switch (symbol.type) {
-                case 'Terminal':
-                  return (
-                    mouseX >= symbol.x_position - SYMBOL_SIZE / 2 &&
-                    mouseX <= symbol.x_position + SYMBOL_SIZE / 2 &&
-                    mouseY >= symbol.y_position - SYMBOL_SIZE / 4 &&
-                    mouseY <= symbol.y_position + SYMBOL_SIZE / 4
-                  );
-                case 'Decision':
-                  return (
-                    mouseX >= symbol.x_position - SYMBOL_SIZE / 1.75 &&
-                    mouseX <= symbol.x_position + SYMBOL_SIZE / 1.75 &&
-                    mouseY >= symbol.y_position - SYMBOL_SIZE / 3 &&
-                    mouseY <= symbol.y_position + SYMBOL_SIZE / 3
-                  );
-                case 'Process':
-                  return (
-                    mouseX >= symbol.x_position - SYMBOL_SIZE / 2 &&
-                    mouseX <= symbol.x_position + SYMBOL_SIZE / 2 &&
-                    mouseY >= symbol.y_position - SYMBOL_SIZE / 4 &&
-                    mouseY <= symbol.y_position + SYMBOL_SIZE / 4
-                  );
-                case 'Data':
-                  return (
-                    mouseX >= symbol.x_position - SYMBOL_SIZE / 2 &&
-                    mouseX <= symbol.x_position + SYMBOL_SIZE / 1.33 &&
-                    mouseY >= symbol.y_position - SYMBOL_SIZE / 4 &&
-                    mouseY <= symbol.y_position + SYMBOL_SIZE / 4
-                  );
-                default:
-                  return false;
-              }
-        });
-
+        const clickedSymbol = symbolsArr.find(symbol => symbolBoundsCheck(symbol, mouseX, mouseY));
 
         if (clickedSymbol) {
-          setDragging(true);
-          setDraggedSymbol(clickedSymbol);
-          setOffsetX(mouseX - clickedSymbol.x_position);
-          setOffsetY(mouseY - clickedSymbol.y_position);
-          setSelectedSymbol(clickedSymbol);
-        } else setSelectedSymbol(null);
+            setDragging(true);
+            setDraggedSymbol(clickedSymbol);
+            setOffsetX(mouseX - clickedSymbol.x_position);
+            setOffsetY(mouseY - clickedSymbol.y_position);
+            setSelectedSymbol(clickedSymbol);
+        }
+        else setSelectedSymbol(null);
 
     };
 
@@ -171,44 +176,12 @@ const GridTool = ({ symbols, onEditSymbol, onDeleteSymbol }) => {
           }
         else {
 
-            const isOverSymbol = symbolsArr.some(symbol => {
-                switch (symbol.type) {
-                    case 'Terminal':
-                    return (
-                        mouseX >= symbol.x_position - SYMBOL_SIZE / 2 &&
-                        mouseX <= symbol.x_position + SYMBOL_SIZE / 2 &&
-                        mouseY >= symbol.y_position - SYMBOL_SIZE / 4 &&
-                        mouseY <= symbol.y_position + SYMBOL_SIZE / 4
-                    );
-                    case 'Decision':
-                    return (
-                        mouseX >= symbol.x_position - SYMBOL_SIZE / 1.75 &&
-                        mouseX <= symbol.x_position + SYMBOL_SIZE / 1.75 &&
-                        mouseY >= symbol.y_position - SYMBOL_SIZE / 3 &&
-                        mouseY <= symbol.y_position + SYMBOL_SIZE / 3
-                    );
-                    case 'Process':
-                    return (
-                        mouseX >= symbol.x_position - SYMBOL_SIZE / 2 &&
-                        mouseX <= symbol.x_position + SYMBOL_SIZE / 2 &&
-                        mouseY >= symbol.y_position - SYMBOL_SIZE / 4 &&
-                        mouseY <= symbol.y_position + SYMBOL_SIZE / 4
-                    );
-                    case 'Data':
-                    return (
-                        mouseX >= symbol.x_position - SYMBOL_SIZE / 2 &&
-                        mouseX <= symbol.x_position + SYMBOL_SIZE / 1.33 &&
-                        mouseY >= symbol.y_position - SYMBOL_SIZE / 4 &&
-                        mouseY <= symbol.y_position + SYMBOL_SIZE / 4
-                    );
-                    default:
-                      return false;
-                  }
-            });
+            const isOverSymbol = symbolsArr.some(symbol => symbolBoundsCheck(symbol, mouseX, mouseY));
 
             document.body.style.cursor = isOverSymbol ? 'grab' : 'auto';
         }
     };
+
 
     const handleMouseUp = () => {
         setDragging(false);
@@ -227,20 +200,20 @@ const GridTool = ({ symbols, onEditSymbol, onDeleteSymbol }) => {
 
     const handleKeyDown = e => {
         // check if the delete key is pressed
-        console.log('deleted')
         if (
-            e.key === 'Delete' ||
+            (e.key === 'Delete' ||
             e.key === 'Backspace' ||
             e.key === 'Del' ||
-            e.key === 'Forward Delete'
+            e.key === 'Forward Delete')
+            && selectedSymbol
         ) {
             // console.log('deleted')
             onDeleteSymbol(selectedSymbol.id);
         }
     };
 
+    // canvas.addEventListener('dblclick', handleDoubleClick);
     canvas.addEventListener('keydown', handleKeyDown);
-
     canvas.addEventListener('mousedown', handleMouseDown);
     canvas.addEventListener('mousemove', handleMouseMove);
     canvas.addEventListener('mouseup', handleMouseUp);
@@ -255,8 +228,10 @@ const GridTool = ({ symbols, onEditSymbol, onDeleteSymbol }) => {
         canvas.removeEventListener('mousemove', handleMouseMove);
         canvas.removeEventListener('mouseup', handleMouseUp);
         canvas.removeEventListener('keydown', handleKeyDown);
+        // canvas.removeEventListener('dblclick', handleDoubleClick);
     };
   }, [dispatch, symbols, dragging, draggedSymbol, selectedSymbol]);
+
 
   // tabindex so deletion via keydown is contained within canvas
   return <canvas id="grid-tool" ref={canvasRef} tabIndex="0"/>;
